BackEnd/recording-service/requirements.txt
Django==5.0.14
django-cors-headers==4.7.0
python-dotenv==1.1.1
mysqlclient==2.2.8
pymongo==4.12.1
redis==5.2.1
celery==5.4.0
boto3==1.38.27
pytz==2025.2
requests==2.32.3
numpy==1.26.4
opencv-python-headless==4.11.0.86
Pillow==11.2.1
livekit==0.13.1
livekit-api==0.8.2
torch==2.7.0
transformers==4.52.4
openai==1.82.1
groq==0.25.0
deep-translator==1.11.4
python-docx==1.1.2
graphviz==0.20.3
pydub==0.25.1
asgiref==3.8.1
sentencepiece
sacremoses